+++ R/DM__Salters_Stracke__2004.R
@@ -1,0 +1,16 @@
+#' Depleted mantle
+#'
+#' A data set containing the estimated composition of the Depleted Mantle as
+#' given by \insertCite{Salters2004;textual}{georefdatar}.
+#'
+#' Please refer to the paper for the units.
+#'
+#' @format A data frame with 1 row and `r ncol(DM__Salters_Stracke__2004)` element concentrations in ppm:\cr
+#' `r paste(names(DM__Salters_Stracke__2004), collapse= ', ')`
+#'
+#' @references {
+#'   \insertRef{Salters2004}{georefdatar}
+#' }
+#'
+#' @importFrom Rdpack reprompt
+"DM__Salters_Stracke__2004"